+++ scripts/import-realmworks.js
@@ -47,7 +47,7 @@ const GS_FLAGS_PIN_PERMISSION = "pinPermission";
 // Register game settings
 //
 Hooks.once('init', () => {
-	
+
 	// See API documentation "ClientSettings"
 	game.settings.register(GS_MODULE_NAME, GS_ASSETS_LOCATION, {
 		name: "Location of Extracted Assets",
@@ -212,7 +212,7 @@ Hooks.once('init', () => {
 // Set up the user interface
 
 Hooks.on("renderSidebarTab", async (app, html) => {
-    if (app.options.id === "compendium") {
+    if (game.user.isGM && app.options.id === "compendium") {
       let button = $("<button class='import-cd'><i class='fas fa-file-import'></i> Realm Works Import</button>")
       button.click(function () {
         new RealmWorksImporter().render(true);
@@ -2271,3 +2271,47 @@ class RealmWorksImporter extends Application
 		if (this.addInboundLinks) delete this.links_in;
 	}
 } // class
+
+//
+// Simple hooks to prevent links being displayed as links if the target of the link isn't OBSERVABLE by the player.
+//
+const _EntityMap = {
+	"JournalEntry" : "journal",
+	"Actor"        : "actors",
+	"RollTable"    : "tables",
+	"Scene"        : "scenes",
+};
+
+async function _checkRenderLinks(sheet, html, data) {	
+	// app  = ActorSheet
+	// html = jQuery
+	// data = object
+	if (game.user.isGM) return;
+	
+	// Original link:
+	//     <a class="entity-link" draggable="true" [ data-entity="JournalEntry" | data-pack="packname" ] data-id=".....">
+	//     <i class="fas fa-th-list">::before</i>
+	//     plain text
+	//     </a>
+	// If the "data-id" isn't observable by the current user, then replace with just "plain text"
+	html.find("a.entity-link").filter( (index,a) => {
+		const dataentity = a.getAttribute('data-entity');	// RollTable, JournalEntry, Actor
+		if (!dataentity) {
+			// Compendium packs are only limited at the PACK level, not an individual document level
+			return game.packs.get(a.getAttribute('data-pack'))?.private;
+		}
+		const entity = _EntityMap[dataentity];
+		if (!entity) {
+			console.warn(`checkRenderLinks#EntityMap does not have '${entity}'`);
+			return false;
+		}
+		const item = game[entity].get(a.getAttribute("data-id"));
+		return !item || !item.testUserPermission(game.user, "OBSERVER");
+	}).replaceWith ( (index,a) => {
+		const pos = a.indexOf("</i> ");
+		return (pos<0) ? a : a.slice(pos+5);
+	});
+}
+
+Hooks.on("renderJournalSheet", _checkRenderLinks);
+Hooks.on("renderActorSheet",   _checkRenderLinks);